Extract core metadata logic into dedicated module
