+++ setuptools/monkey.py
@@ -99,11 +99,11 @@ def patch_all():
 
 
 def _patch_distribution_metadata():
-    from . import dist
+    from . import _core_metadata
 
     """Patch write_pkg_file and read_pkg_file for higher metadata standards"""
     for attr in ('write_pkg_file', 'read_pkg_file', 'get_metadata_version'):
-        new_val = getattr(dist, attr)
+        new_val = getattr(_core_metadata, attr)
         setattr(distutils.dist.DistributionMetadata, attr, new_val)
